remove flow/connectors/postgres/qrep_sql_sync.sql
